Add Dockerfile
Add a Dockerfile to support Docker-based development of the website.

The Dockerfile can be used to create a Docker image for the simple-icons
website - either to build & serve it or for full development.

To be able to both build continuosly (npm run build:watch) and serve
(npm run serve), a new npm run command is introduced - called npm run
serve:dev - that runs the two other commands in parallel using
npm-run-all's [1] `run-p` utility.

The Contributing Guidelines have been updated to include a section on
how to use Docker for development. The table of development commands has
also been updated to include the new `npm run serve:dev` command (as
well as the previously undocumented `format` and `lint` commands).

The Docker image is pretty straightforward, following best practices
such as copying manifest files and installing dependencies before the
full project is copied. There are two less common things in the
Dockerfile:

- An environment variable is set to prevent Puppeteer from installing
  Chromium. Although it is possible to support running Puppeteer with
  Chromium in an Alpine Docker container, I choose to omit it for now
  for simplicity.
- The `xdg-utils` system dependency is installed, this appears to be
necessary in order to be able to run anywhere [2].

--
1. https://www.npmjs.com/package/npm-run-all
2. https://www.npmjs.com/package/anywhere