Update Contributing Guidelines
Update the Docker example commands for usability.

The older `docker run` example required that you have Node/NPM locally
and had run `npm install`. The new example allows you to run the
webserver in Docker, staying in sync with the changes to the website,
without requiring NPM or Node locally.