Fix style issue with
Fix a stylistic issue where the ".copied" class on the SVG copy button
would not properly display, in dark mode, after unhovering. This is due
to the fact that the z-index: 1; styling is only applied on hover (if
hovering is available on the system). This has been fixed by adding
z-index: 1; to the .copied (and :focus) selector as well.