Move sticky header shadow down to prevent/reduce visual bug
Prevent/reduce the visual bug where the header shadow appears above the
website header. This bug was found in Chromium-based browsers and
appears to be caused by slower rendering updates, in particular due to
the large amount of elements on the page.

The problem was "fixed" by having the starting position of the shadow's
element be lower on the page, as a result it won't move up as far on the
page, preventing (or at least reducing the precense) of the bug.