Add content-visibility to grid items
Add the new content-visibility CSS property [1] to grid items and use
content-intrinsic-size [2] to define a "standard" height, before the
grid item is rendered. This is currently set to 11.25rem (or 180px at a
root font-size of 16px, which seems to be the default of browsers with
the existing CSS) which is roughly the height of a grid item. Small grid
items are ~170px (short title, no guidelines/license), medium grid items
are ~190px (long title, no guidelines/license). Some large grid items,
~220px (short title, guidelines and license), exist but they're not very
common.

--
1. https://developer.mozilla.org/en-US/docs/Web/CSS/content-visibility
2. https://developer.mozilla.org/en-US/docs/Web/CSS/contain-intrinsic-size