Use filesystem cache for webpack and cache that in CI (#1)
This has a couple of advantages over the cache-hit based build solution:

- This cache is also useful to the publish workflow.
- This cache is useful accross CI runs.
- This cache is also useful locally.
- This is (arguably) more intuitive.

The downside being that this is probably a bit less performent.

Note that I put the cache for webpack in .cache/webpack, that way the
cache may be used for other purposes as well (e.g. jest test caching).