Improve HTML (#83)
* HTML improvements: rel & role attributes

- Add `role="button"` to button-like <a> elements.
- Remove unnecessary `rel="noopener"` if `target="_blank"` is not used.
- Remove unnecessary role attribute on `<header>`.
- Remove unnecessary role attribute on `<main>`.
- Remove unnecessary role attribute on `<footer>.

* Fix use of aria-hidden, always set to `"true"`

Always set the value of aria-hidden explicitly to `"true"`, as
`aria-hidden`/`aria-hidden=""` is not correct [1].

Also updated a related test to verify that dynamically hidden elements
are properly assigned the `aria-hidden` attribute.

--
1. https://stackoverflow.com/a/55061234

* Add aria-labelledby to control button SVGs

* Add license link to `<head>`

Add a license link to the website header, and include LICENSE.md
in the build. I chose to rename the license to 'license.txt" in the
build to prevent browsers from trying to download the .md file.

* Test license file availability