Remove JavaScript blurs (#126)
* Remove blurring of buttons on click

This resets the tab index to the top of the page, which is very
frustating for keyboard users. The git history doesn't provide a clue
why these calls to blur are here (perhaps it's taken straight from the
old website), but I tested these changes and everything seems fine.

* Remove blur for 'change' listener of search input

This blur doesn't seem to do anything. The input is already out of focus
when the event fires.