Alleviate package-lock changing problems (#108)
* Set allowed npm version in manifest's engines

- Fix the name of the engines field
- Specify the suported npm version range

* Set `eninge-instrict=true` in .npmrc

This will force anyone using npm commands in this project to adhere to
the values of the "engines" field of this project (and dependencies)
(unless they use --force).

* Add an `.nvmrc` file

To make it easier for contributors that use NVM to switch to the right
NodeJS & npm version.

* Set lockfile-version=1 in .npmrc

This tells npm version 8.1.0 and higher to use lockfile-version 1
rather than 2. In practice, it seems 8.1.0 and 8.1.1 have some problems
recognizing the option when it is specified in the .npmrc file, hence I
allowed npm versions higher than 8.1.2 to run `npm install` (in addition
to <7.0.0).